+++ src/cli/commands/delete.ts
