+++ src/cli/commands/pack.ts
